feat: :fire: Rename parckage doc to docs
