+++ docs/assests/MokapsVisualizingGitPart1.png
