+++ docs/assests/MokapsVisualizingGitPart3.png
