+++ docs/assests/SocketRoom.png
