+++ docs/assests/docker-architecture.webp
